moved the provider balance tracker to a separate provider package
