+++ cmd/di.go
@@ -54,6 +54,7 @@ import (
 	"github.com/mysteriumnetwork/node/services/openvpn/discovery/dto"
 	"github.com/mysteriumnetwork/node/session"
 	"github.com/mysteriumnetwork/node/session/balance"
+	balance_provider "github.com/mysteriumnetwork/node/session/balance/provider"
 	session_payment "github.com/mysteriumnetwork/node/session/payment"
 	payment_factory "github.com/mysteriumnetwork/node/session/payment/factory"
 	payments_noop "github.com/mysteriumnetwork/node/session/payment/noop"
@@ -316,7 +317,7 @@ func newSessionManagerFactory(
 			}
 
 			// TODO: the ints and times here need to be passed in as well, or defined as constants
-			tracker := balance.NewProviderBalanceTracker(&timeTracker, amountCalc, time.Second*5, 0)
+			tracker := balance_provider.NewBalanceTracker(&timeTracker, amountCalc, time.Second*5, 0)
 			validator := validators.NewIssuedPromiseValidator(consumer, provider, issuer)
 			return session_payment.NewSessionBalance(sender, tracker, promiseChan, time.Second*5, time.Second*1, validator), nil
 		}
